Refactored out db connection to separate module
